WAV reader: read >4 GiB files
Has data chunk size set to UINT32_MAX and size is deduced from file
length. This is consistent to what FFmpeg (MediaInfo, MPV and perhaps
others) does.

+++ src/audio/wav_reader.h
@@ -50,8 +50,8 @@ struct wav_metadata {
         uint16_t valid_bits;
         uint32_t channel_mask;
 
-        unsigned int data_size;
-        unsigned int data_offset; // from the beginning of file
+        long long data_size;
+        long long data_offset; // from the beginning of file
 };
 
 #define WAV_HDR_PARSE_OK           0
